Switch to a pinned image instead of building the Dockerfile
This allows us to get around the bitrotted notebook environment and avoid
blocking interns. We will eventually have to upgrade to the latest version of
everything, but can defer that until June when we will have more intern support.

This is a workaround for https://github.com/e-mission/em-public-dashboard/issues/37