Add tentative support for the "air" mode
Not merging this to main yet because air then dwarfs everything else in terms
of emissions. We already show this as "other" on the pie chart, but we ignore
other trips for emission calulations, making them look lower than they actually
are.

This change reintroduces them, but then all other emissions are dwarfed in the
timeline.

Need to think about how best to represent this.