Make the first date be non-hardcoded
This also required some refactoring of the `getMonthList` function.

Without the refactoring, we would get into an infinite loop where we kept
iterating over dates forever. I am not 100% sure if this is because the current
month = 9 and the start month for the selected program was also 9, or whether
it was because the start_month was a string.

+ convert the configured values into ints otherwise you end up with a month
that is 91111111

This works now