Ensure that the plots are actually loaded correctly
Before this change, the "value" in the option was misformatted, AND did not
have the initial zero padding for the month.

Fixes are to create a padded value for the month
(https://bobbyhadz.com/blog/javascript-pad-number-with-leading-zeros)
and then use built-in javascript templates (back ticks) to construct the
options in a more legible fashion.
https://stackoverflow.com/questions/610406/javascript-equivalent-to-printf-string-format
https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Template_literals

So now we have a solution for templating in raw javascript that does not depend
on any templating libraries.  Yay!