Ensure that the page title is also dynamic
NREL OpenPATH dashboard makes it seem like this is *the* dashboard for NREL
OpenPATH. But in fact, it is the dashboard for each individual study or
program. So let's actually populate the title (page title + top visible title)
with the study-specific information.

As a bonus, this is an example for @sebastianbarry on how to configure text
without having to include a templating library.