Infer the study config from the URL on the frontend
Javascript browsers cannot, for security reasons, read environment variables.

There are workarounds using config files that are even automatically generated
using webpack, but we are not using webpack now and are unlikely to use it in
the future given that this is a temporary solution.

So we use the same approach as we do on the phone to infer the study name from
the URL.

However, we cannot test this functionality easily on dev, because the hostname
is always "localhost". Looking at search params only in that case to finish testing.

Testing done:
- Accessed frontend using http://localhost:3274/, saw the staging config
- Accessed it using http://localhost:3274/?study_config=durham, saw the durham config