Handle retrieval of stage data better
We use the study name retrieved from the connect URL to retrieve the
configuration from the server. For the stage environment, this study name is
"stage". But there are separate stage study and program configurations. So we
expand the hack to work with localhost to stage as well. In the hack, we put
the config as a URL parameter and fall back to "stage-program" if it is not present.

Testing done:
- loaded localhost with and without the query parameter
    - with query parameter, loaded durham
    - without query parameter, loaded stage-program