Handle cold start with missing replaced mode properly
Instead of passing in a hardcoded value for the labels_per_trip depending on
the study_type, automatically infer it from the columns provided.

Even if the study is a program, ensure that the replaced mode is only mapped
and replaced when it exists. Otherwise, skip with a warning

Testing done:
On the same dataset with cold start issues:
https://github.com/e-mission/em-public-dashboard/issues/65
loaded the data correctly and generated all graphs other than the commute mode
share correctly