Fix the quality text for < 10 miles as well
We plot all labeled trips that are < 10 miles, so we want to fix the n for it
as well. It is unclear what the universe is here - all labeled trips, or all
trips < 10 miles, but let's go with all labeled trips to be consistent with the
other quality text.

This is a fix similar to
https://github.com/e-mission/em-public-dashboard/pull/66/commits/3dabe72079dbb0a345ad44ad7f8c72eddd519e04
and also addresses an issue similar to
https://github.com/e-mission/em-public-dashboard/issues/65#issuecomment-1253075694