Force the missing plot size to 10x10 + avoid trip_label_count copy/paste
We originally tried to make the figure for the missing subplots depend on the
kind of plot. However, it turns out that non-square figsizes lead to bad
overlaps (https://github.com/e-mission/em-public-dashboard/pull/70#issuecomment-1257124459)

Forcing the size to 10x10 works even when displayed in a rectangular grid tile
https://github.com/e-mission/em-public-dashboard/pull/70#issuecomment-1257121306
which is why we had set all the kind = pie

This means that the parameter is no longer useful, so we remove it all

Finally, we were copy-pasting the equivalent of `trip_label_count` in the < 10
miles metric. We avoid this by moving it outside the function (which
necessitates passing in expanded_ct throughout instead of having it be a
default parameter) and then invoking it from the notebook.