Add basic spatial visualization
- Start/end points
- Heatmap
- Trajectory mapping
- Pixel-specific ratios

Most of these analyses were generated for Denver as part of the exploration of MEP integration
The trajectory mapping was initially generated for Durham

The notebook relies on two sets of shapefiles:
- Denver boundary
- Denver boundary, gridded into 100m pixels

Both of these are courtesy the MEP team

Note that the notebook currently takes several hours to run. The majority of
this time is spent in the trajectory mapping, which takes ~ 5 hours on my very powerful laptop.