Added cert.sh + Modified Dockerfiles
1. docker/cert.sh
- Uses new environment variable PROD_STAGE variable to determine whether in staging / production environment and only then install certificates, else skip.

2. Dockerfile
Added environment variables here, keeping the same default ENV values as the ones in docker-compose.yml.
Not adding in docker/Dockerfile.dev, since this change is being done primarily with the objective to aid with the automated build and push to Dockerhub.
This pushed image would then be the one that would be used in the internal environments as the base image, which would be based on the non-dev Dockerfile.