Removing ENV variables from Dockerfile
The other ENV vars that were added in Docker-compose files are currently present in the Dockerfile in the external repo root.
Removing them from there and can add them as per requirement.
For local testing, add them when “docker run” is used using the -e flag.
For usage in stage / production, can be set by cloud team in AWS Codebuild as they currently do.

CRON_MODE
CRON_MODE can be moved to setting when docker run command is executed using -e flag.
This is because this is required in start_notebook.sh in CMD layer in Dockerfile which is executed when the container is run.
Hence, CRON_MODE value can be supplied later and not needed to be embedded in image.

PROD_STAGE
Not setting PROD_TRUE in docker run command since it is required during docker image build to decide whether to add certificates or not.
Hence, adding it in the Dockerfile so it’s available during docker build.