🔥 Remove the code to preprocess cutoffs
The code is only called in this notebook.
It was called twice to generate separate cutoffs for confirmed and sensed trips
separately, but I am not convinced that separate cutoffs actually make sense
for two bars in the same chart. It is hard to compare two values when they
don't have the same base.

And the cutoff here is really just a data-driven shortcut for "short trips"
If we had hardcoded it to 10 miles, we would have looked at trips less than 10
miles in both datasets. If we want to determine which trips are short, we
should look at all trips.
https://github.com/e-mission/em-public-dashboard/pull/123#discussion_r1586845105