♻️  Final refactor: pass in the aggregation type
So that we can support both number of trips and trip distance.
The aggregation type is now either `count` (for number of trips) or `sum` (for
trip distance). The count functionality has been switched from the previous
`value_counts` to using the standard `group_by` and `agg`

Testing done:
- generic_metrics runs through with no errors
https://github.com/e-mission/em-public-dashboard/pull/123#discussion_r1588038687
https://github.com/e-mission/em-public-dashboard/pull/123#discussion_r1588097070

This allows us to remove the old `process_distance_data` from `plots.py`
And we are now able to convert the distance metrics to also use this new format