🐛 Fix the generation of alt text and HTML for a single bar chart
When we generate a figure with a single bar chart (aka nRows = 1, we get a single `ax` instead of an axarray.
In similar fashion, if there is a single bar charts, I created text_results as
a single list instead of a list of lists.

But then, while using the `text_results`, I hardcoded accesses to
`text_results[0][0]` etc which should have returned the string for the first
plot but ended up returning the first character of the string.

Handled this by:
- Iterating over the number of axes to concatenate the strings instead of assuming that there were two
- If there was only one axis, wrapping the text_result in a list so that it was
  a list of one list and so consistent with the expected data format.

This fixes https://github.com/e-mission/em-public-dashboard/pull/123#issuecomment-2091747733
consistent with https://github.com/e-mission/em-public-dashboard/pull/123#issuecomment-2092028278

Testing done:
After the changes, the HTML for the purpose chart (single bar) was as below

```
    <!DOCTYPE html>
    <html>
    <body>
        <p>Number of trips for each purpose (selected by users)
Based on 1417 confirmed trips from 11 users
of 2728 total  trips from 13 users (51.94%)</p>
    <p>
<tr><td>Other</td><td>372</td><td>26.3%</td></tr><tr><td>Home</td><td>339</td><td>23.9%</td></tr><tr><td>Recreation/Exercise</td><td>169</td><td>11.9%</td></tr><tr><td>Shopping</td><td>128</td><td>9.0%</td></tr><tr><td>Meal</td><td>111</td><td>7.8%</td></tr><tr><td>Entertainment/Social</td><td>94</td><td>6.6%</td></tr><tr><td>Personal/Medical</td><td>93</td><td>6.6%</td></tr><tr><td>Work</td><td>56</td><td>4.0%</td></tr><tr><td>Transit transfer</td><td>48</td><td>3.4%</td></tr><tr><td>School</td><td>4</td><td>0.3%</td></tr><tr><td>not_a_trip</td><td>2</td><td>0.1%</td></tr><tr><td>Religious</td><td>1</td><td>0.1%</td></tr></p>
    </body>
    </html>
```