💄 💩 Fix legend display
A temporary fix for now is to pin the bbox to the bottom and not the top.

I think that the real fix is to use ncols, because then it will work for both
sensed and labeled legends. What if we have more than 5 sensed values?

But we will need to revisit this anyway when we have inferred modes, and when
we display surveys, and when we use pandas barh, so punting on this and
implementing the temporary fix for now

Testing done:
https://github.com/e-mission/em-public-dashboard/pull/123#issuecomment-2094563995