🐛 Fix regression while generating sensed quality text with testers and participants
```
def get_quality_text_sensed(df, cutoff_text="", include_test_users=False):
```

now has a `cutoff_text` argument. However, the invocation from scaffolding was still

```
    quality_text = get_quality_text_sensed(expanded_ct, include_test_users)
```

So the boolean `True` was interpreted as the cutoff_text and so the text that was generated was "Based on 2728 trips (True) from 13 users".

After fixing that, this option seems to work.

Testing done:
https://github.com/e-mission/em-public-dashboard/pull/123#issuecomment-2094592574