Use the `ble_sensed_mode` for fleet operations
- Comment in the previous code for using the `ble_sensed_mode`
- Fixed color issues; we cannot use `colors_sensed` any more since the
  `ble_sensed_mode` will not be one of the 5 hardcoded values. So we find the
set of unique values using a `groupby+agg` and then create a fake dictionary of
the unique keys mapped to themselves to pass in. Note that the returned value
will be `colors_mode` while we need to use `colors_sensed` in the graphs to
handle the non-fleet case. So for the fleet case, we set `colors_sensed` to
`color_mode`

Now that I type that out, I see that the better fix would be to have a
`colors_sel` that is set appropriately for fleet/non-fleet and then pass in
`colors_sel` instead of `colors_mode`. But that would be a bigger change, so
will leave it for @abby-wheelis to handle in the next round of changes

Testing done:
https://github.com/e-mission/em-public-dashboard/pull/135#issuecomment-2102004470