fix error on older configs that don't have survey_info
The survey_info field is not present on configs older than approx. late 2022 / early 2023.
This caused an error on the public dashboard webpage which prevented the dropdown menus from populating or any charts from showing. Observed on smart-commute-ebike in production.

This fix considers that survey_info might not be present and and only proceeds if it is. Tested locally with the smart-commute-ebike config and observed that the error goes away when the fix is applied